[blog] Embed RicosViewer inside PostContent (#290)
